Learning machine learning from a book
Just trying to get the desired output from this code  which is three lines of numbers in the array. Instead its throwing up every error message under the sun which is confusing the shit out of me I even bought a python book to familiarize myself with the language more but its doesn't cover mathematical librarys, well statistical libraries and machine learning etc.